auth/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.2
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.12
GeoAlchemy2==0.17.1
greenlet==3.1.1
h11==0.14.0
humanfriendly==10.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
pyreadline3==3.5.4
python-dotenv==1.1.0
python-http-client==3.3.7
python-multipart==0.0.20
sendgrid==6.11.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starkbank-ecdsa==2.2.0
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
